DEV: Restructure `encounter` module.
