--- celest/__init__.py
@@ -1,4 +1,0 @@
-from .satellite import Satellite
-from .groundposition import GroundPosition
-from .encounter import Encounter, EncounterSpec
-from .astronomy import CelestialObject
